refactor: move pagination controls component
