+++ src/components/main/pagination-controls/pagination-controls.css
